+++ packages/ui/src/components/PaginatedCarousel/PaginatedCarouselRoot/brands/index.ts
@@ -1,0 +1,6 @@
+import { type BrandVariants } from '@/lib/brand-registry';
+import { paginatedCarouselRootAcornVariants } from './PaginatedCarouselRoot.acorn.brand';
+
+export const paginatedCarouselRootBrandVariants = {
+  acorn: paginatedCarouselRootAcornVariants
+} as const satisfies BrandVariants<typeof paginatedCarouselRootAcornVariants>;