+++ packages/ui/src/components/PaginatedCarousel/PaginatedCarouselSlide/brands/index.ts
@@ -1,0 +1,6 @@
+import { type BrandVariants } from '@/lib/brand-registry';
+import { paginatedCarouselSlideAcornVariants } from './PaginatedCarouselSlide.acorn.brand';
+
+export const paginatedCarouselSlideBrandVariants = {
+  acorn: paginatedCarouselSlideAcornVariants
+} as const satisfies BrandVariants<typeof paginatedCarouselSlideAcornVariants>;