+++ packages/ui/src/components/PaginatedCarousel/PaginatedCarouselViewport/brands/index.ts
@@ -1,0 +1,8 @@
+import { type BrandVariants } from '@/lib/brand-registry';
+import { paginatedCarouselViewportAcornVariants } from './PaginatedCarouselViewport.acorn.brand';
+
+export const paginatedCarouselViewportBrandVariants = {
+  acorn: paginatedCarouselViewportAcornVariants
+} as const satisfies BrandVariants<
+  typeof paginatedCarouselViewportAcornVariants
+>;