+++ packages/tokens/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @perimetre/tokens
 
+## 0.4.1
+
+### Patch Changes
+
+- 5e03749: Adds more microbird colors
+
+## 0.4.0
+
+### Minor Changes
+
+- a3ab7d8: Changing p tag into span tag in eyebrow field
+- a3ab7d8: Created SectionHorizontalHeader component
+
 ## 0.3.0
 
 ### Minor Changes